+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+1.2.1
+=====
+
+* Add xunit and clover report file options
+
+1.2.0
+=====
+
+* Adds the ability to test Symfony commands (see `atoum\AtoumBundle\Test\Units\CommandTestCase`)
+
 1.1.0
-=======
+=====
 
   * Add command to launch tests on bundles.
   * Add fluent interface for controllers testing